Move `printSchema()` compatibility code to a module
Clarifies what is part of the API schema and what is part of graphql-js
compatibility.
